use the Interrupt Stack Table for all interrupts
In traditional OSes, code that executes in ring 0 is compiled without
stack red zones in order to avoid stack corruption caused by interrupts
which push data on the stack. We can ask for developers to compile their
Dune application with -mno-red-zone, but they must also compile all
their depending libraries with the same flag. Since this is not
practical, we enable an IA-32e feature that switches the stack pointer
for all the interrupts.